+++ src/common/utf8.rs
@@ -16,6 +16,9 @@ pub trait Utf8Encoding<'a>: private::Sealed {
     /// Represents the type of component that will be derived by this encoding
     type Components: Utf8Components<'a>;
 
+    /// Static label representing encoding type
+    fn label() -> &'static str;
+
     /// Produces an iterator of [`Utf8Component`]s over the given the byte slice (`path`)
     fn components(path: &'a str) -> Self::Components;